requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
joblib==1.3.2
requests==2.31.0
llama-cpp-python==0.2.11
diffusers==0.24.0
transformers==4.35.0
torch==2.1.0
accelerate==0.24.0 